==> notebooks/first_experiments/03_pdf_preperation.ipynb <==
# PDF papers preperation

import json

# Read the content of the JSON file into a dictionary
with open('../data/articles_info.json', 'r') as file:
    articles_info = json.load(file)

print(articles_info)

### Match PDF names to saved article titles and their IDs

import os

# Initialize counters and list to store unmatched titles
match_count = 0
no_match_count = 0
unmatched_titles = []
matched_ids = {}

# Function to check if a title matches any title in the articles_info dictionary
def title_matches(title, articles_info):
    matches = set()
    for category, articles in articles_info.items():
        for article_id, article in articles.items():
            if 'title' in article:
                if title.lower() in article['title'].replace(':', '').lower() or article['title'].replace(':', '').lower() in title.lower():
                    matches.add(article_id)
                if title.lower()[:50] in article['title'].replace(':', '').lower() or title.lower()[10:60] in article['title'].replace(':', '').lower():
                    matches.add(article_id)
    return list(matches)

# Iterate through all PDF files in the directory
for root, dirs, files in os.walk(main_dir):
    for file in files:
        if file.endswith('.pdf'):
            file_path = os.path.join(root, file)
            file_name = os.path.splitext(file)[0]  # Get the file name without the extension
            match_ids = title_matches(file_name, articles_info)
            if match_ids:
                match_count += 1
                matched_ids[file_name] = match_ids
            else:
                no_match_count += 1
                unmatched_titles.append(file_name)

print(f"Match count: {match_count}")
print(f"No match count: {no_match_count}")
print("Unmatched titles:")
for title in unmatched_titles:
    print(title)


multiple_matches = {key: value for key, value in matched_ids.items() if len(value) > 1}
print(multiple_matches)

### Append IDs to PDF names

# Iterate through all PDF files in the directory
# for root, dirs, files in os.walk(main_dir):
#     for file in files:
#         if file.endswith('.pdf'):
#             file_path = os.path.join(root, file)
#             file_name = os.path.splitext(file)[0]  # Get the file name without the extension
#             if file_name in matched_ids:
#                 match_id = matched_ids[file_name][0]  # Get the first matched ID
#                 new_file_name = f"{match_id}_{file}"
#                 new_file_path = os.path.join(root, new_file_name)
#                 os.rename(file_path, new_file_path)
#                 print(f"Renamed: {file} to {new_file_name}")
#             else:
#                 print(f"No match for: {file}")
