==> notebooks/first_experiments/05_matching_chunks_identification.ipynb <==
# Identify the 3 top matching chunks using LlamaIndex

LlamaIndex:
- Github: https://github.com/run-llama/llama_index
- Documentation: https://docs.llamaindex.ai/en/latest/

import pandas as pd

annotated = True

# read the xlsx data into a pandas dataframe
df = pd.read_excel(f"../data/ReferenceErrorDetection_data{'_annotated' if annotated else ''}.xlsx")

df.head()

## Get text from reference article

grobid_model = "full_model"
extension = "xml"

import glob

def get_file_path(reference_article_id):
    # Construct the file path pattern using the Reference Article ID of the first entry
    file_pattern = f"../data/extractions/{grobid_model}/{reference_article_id}*.{extension}"

    # Find the file that matches the pattern
    file_list = glob.glob(file_pattern)
    if file_list:
        file_path = file_list[0]
        return file_path
    else: 
        print("No matching file found.")
        return None

import xml.etree.ElementTree as ET

def get_reference_text(reference_article_id):
    global extension
    
    # Get the file path
    file_path = get_file_path(reference_article_id)
    
    if file_path:
        if extension == "txt":
            # Read the text file
            with open(file_path, 'r') as file:
                reference_text = file.read()
            return reference_text

        elif extension == "xml":
            # Parse the XML file
            tree = ET.parse(file_path)
            root = tree.getroot()

            # Extract the text content from the XML file
            reference_text = ''.join(root.itertext())
            return reference_text

## Set OpenAI key

# Read the content of open_ai_key.txt into a variable
with open('../open_ai_key.txt', 'r') as file:
    open_ai_key = file.read().strip()

## Setting up vector index

### Reloading or generating index

embedding = "te3l"

if embedding == "te3l":
    model_embeddings = "text-embedding-3-large"
elif embedding == "te3s":
    model_embeddings = "text-embedding-3-small"

from llama_index.embeddings.openai import OpenAIEmbedding
from llama_index.core import Settings

Settings.embed_model = OpenAIEmbedding(model=model_embeddings, api_key=open_ai_key)

from llama_index.core import VectorStoreIndex, Document
from llama_index.embeddings.openai import OpenAIEmbedding
from llama_index.core.node_parser import SentenceSplitter
from llama_index.core.ingestion import IngestionPipeline

def create_index(reference_text, chunk_size, chunk_overlap):
    # create the pipeline with transformations
    pipeline = IngestionPipeline(
        transformations=[
            SentenceSplitter(chunk_size=chunk_size, chunk_overlap=chunk_overlap),
            OpenAIEmbedding(model=model_embeddings, api_key=open_ai_key)
        ]
    )

    # run the pipeline
    nodes = pipeline.run(documents=[Document(text=reference_text)])
    index = VectorStoreIndex(nodes)
    return index

from llama_index.core.node_parser import TokenTextSplitter

def create_chunks(text, chunk_size=256, chunk_overlap=20):
    token_text_splitter = TokenTextSplitter(chunk_size=chunk_size, chunk_overlap=chunk_overlap)
    chunks = token_text_splitter.split_text(text)
    return chunks

def create_index_with_chunking_first(reference_text, chunk_size, chunk_overlap):
    reference_chunks = create_chunks(reference_text, chunk_size, chunk_overlap)
    documents = [Document(text=chunk) for chunk in reference_chunks]
    index = VectorStoreIndex.from_documents(documents)
    return index

from llama_index.core import StorageContext, load_index_from_storage
import os

def load_or_create_index(article_id, reference_text, no_prev_chunking, chunk_size=256, chunk_overlap=20, only_checking=False):
    index_path = f"../data/vector_indices{'/annotated_data' if annotated else ''}/{embedding}{'_no_prev_chunking' if no_prev_chunking else ''}/{grobid_model}/{article_id}/"
    index = None
    if only_checking:
        if os.path.exists(index_path) and os.listdir(index_path):
            print(article_id + ": Index exists.")
            return True

    try:
        storage_context = StorageContext.from_defaults(persist_dir=index_path)
        index = load_index_from_storage(storage_context)
        print(article_id + ": Loaded existing index.")
    except Exception as e:
        print(e)
        print(article_id + ": Creating a new index.")
        try: 
            if no_prev_chunking:
                index = create_index(reference_text, chunk_size, chunk_overlap)
            else:
                index = create_index_with_chunking_first(reference_text, chunk_size, chunk_overlap)
            index.storage_context.persist(persist_dir=index_path)
        except Exception as e:
            print(e)
            print(article_id + ": Failed to create index.")
            print(reference_text)
    return index

### Create Indices for all reference articles

%%time

ids_to_index = []
no_prev_chunking = True

for _, row in df.iterrows():
    if row['Reference Article Downloaded'] == 'Yes':
        reference_article_id = row['Reference Article ID']
        if reference_article_id and (len(ids_to_index) == 0 or reference_article_id in ids_to_index):
            reference_text = get_reference_text(reference_article_id)
            index = load_or_create_index(reference_article_id, reference_text, no_prev_chunking, only_checking=True)

## Identifying top 3 chunks

from llama_index.core.retrievers import VectorIndexRetriever

def get_top_k_similar_chunks(statement, index, k=3):
    retriever = VectorIndexRetriever(
        index=index,
        similarity_top_k=k,
    )
    retrieved_nodes = retriever.retrieve(statement)
    return retrieved_nodes

import json
import os

def save_similar_chunks(doc_ids, reference_id, corrected_statements):
    file_path = f"../data/similar_chunks{'/annotated_data' if annotated else ''}{'/corrected_statements' if corrected_statements else ''}/{embedding}{'_no_prev_chunking' if no_prev_chunking else ''}/{grobid_model}/{reference_id}.json"
    os.makedirs(os.path.dirname(file_path), exist_ok=True)
    with open(file_path, 'w') as file:
        json.dump(doc_ids, file)

def load_similar_chunks(reference_id, corrected_statements):
    file_path = f"../data/similar_chunks{'/annotated_data' if annotated else ''}{'/corrected_statements' if corrected_statements else ''}/{embedding}{'_no_prev_chunking' if no_prev_chunking else ''}/{grobid_model}/{reference_id}.json"
    with open(file_path, 'r') as file:
        doc_ids = json.load(file)
    return doc_ids

### Save the document ids and text contents of the top k chunks of all reference articles to the df

def get_doc_ids(response):
    doc_ids = []
    for node in response:
        doc_ids.append(node.dict()['node']['id_'])
    return doc_ids

def save_top_k_chunk_ids(df, no_prev_chunking, corrected_statements=False, k=3):
    for _, row in df.iterrows():
        if row['Reference Article Downloaded'] == 'Yes':
            reference_article_id = row['Reference Article ID']
            print(f"------ Starting {reference_article_id} ------")
            
            # Try to load similar chunks first
            try:
                doc_ids = load_similar_chunks(reference_article_id, corrected_statements)
                print("Loaded similar chunks successfully.")
            except FileNotFoundError:
                # Load reference text and create chunks
                reference_text = get_reference_text(reference_article_id)
                
                # Load or create index
                index = load_or_create_index(reference_article_id, reference_text, no_prev_chunking)
                
                # Get the statement and retrieve top chunks
                statement = row["Statement with Citation"]
                if corrected_statements:
                    statement = row["Corrected Statement"]
                    if pd.isna(statement) or statement == "null":
                        statement = row["Statement with Citation"]
                    else:
                        print(statement)

                print("Receiving top chunks")

                try:
                    response = get_top_k_similar_chunks(statement, index, k)
                    doc_ids = get_doc_ids(response)
                    
                    # Save the top chunks
                    print("Saving top chunks")
                    save_similar_chunks(doc_ids, reference_article_id, corrected_statements)
                except Exception as e:
                    print(e)
                    print("Failed to get top chunks.")
            print("")

%%time 

corrected_statements = False

save_top_k_chunk_ids(df, no_prev_chunking, corrected_statements=corrected_statements)

import pandas as pd

output_dir = f"../data/dfs{'/annotated_data' if annotated else ''}{'/corrected_statements' if corrected_statements else ''}/{embedding}{'_no_prev_chunking' if no_prev_chunking else ''}/{grobid_model}/"
os.makedirs(output_dir, exist_ok=True)
# df2 = pd.read_pickle(os.path.join(output_dir, f"ReferenceErrorDetection_data_with_chunk_info.pkl"))

def add_top_k_chunk_ids_and_texts_to_df(df, corrected_statements=False, k=3):
    doc_ids_list = []
    doc_texts_list = []
    for _, row in df.iterrows():
        if row['Reference Article Downloaded'] == 'Yes':
            reference_article_id = row['Reference Article ID']

            if len(ids_to_index) != 0 and reference_article_id not in ids_to_index:
                doc_ids_list.append(row[f'Top_{k}_Chunk_IDs'])
                doc_texts_list.append(row[f'Top_{k}_Chunk_Texts'])
                continue
            
            print(f"------ Starting {reference_article_id} ------")

            # load index
            index_path = f"../data/vector_indices{'/annotated_data' if annotated else ''}/{embedding}{'_no_prev_chunking' if no_prev_chunking else ''}/{grobid_model}/{reference_article_id}/"
            storage_context = StorageContext.from_defaults(persist_dir=index_path)
            index = load_index_from_storage(storage_context)

            # load similar chunks
            doc_ids = load_similar_chunks(reference_article_id, corrected_statements)
            doc_texts = [index.docstore.docs[doc_id].text for doc_id in doc_ids]

            # add to lists
            doc_ids_list.append(doc_ids)
            doc_texts_list.append(doc_texts)
        else:
            doc_ids_list.append(None)
            doc_texts_list.append(None)
    
    df[f'Top_{k}_Chunk_IDs'] = doc_ids_list
    df[f'Top_{k}_Chunk_Texts'] = doc_texts_list
    return df

ids_to_index = []
corrected_statements = False
df2 = add_top_k_chunk_ids_and_texts_to_df(df, corrected_statements=corrected_statements)

df2.head()

import os

# Ensure the directory exists
output_dir = f"../data/dfs{'/annotated_data' if annotated else ''}{'/corrected_statements' if corrected_statements else ''}/{embedding}{'_no_prev_chunking' if no_prev_chunking else ''}/{grobid_model}/"
os.makedirs(output_dir, exist_ok=True)

# Save the DataFrame to a pickle file
df2.to_pickle(os.path.join(output_dir, f"ReferenceErrorDetection_data_with_chunk_info.pkl"))
