==> notebooks/first_experiments/06_2_model_prompting_local.ipynb <==
# Model prompting with installed Ollama models

%%capture output
%run ./06_0_prompt_creation.ipynb

import pandas as pd

embedding = "te3l" # / "te3s"
grobid_model = "full_model"
no_prev_chunking = True

annotated = True
corrected_statements = True

path = f"../data/dfs{'/annotated_data' if annotated else ''}{'/corrected_statements' if corrected_statements else ''}/{embedding}{'_no_prev_chunking' if no_prev_chunking else ''}/{grobid_model}/ReferenceErrorDetection_data_with_chunk_info.pkl"
print(path)

# read the dataframe from a pickle file
df = pd.read_pickle(path)

df[df['Corrected Statement'].notna()]

import ollama

def send_prompt(prompt, model):
    response = ollama.chat(model=model, messages=[{'role': 'user', 'content': prompt}])
    return response['message']['content']

import time

def prompting_model(df, model, skip_existing=False, save_intermediate_results=False, save_every=10, two_labels=False):
    print(f"Prompting model: {model}")

    # Create a new column in the dataframe to store the responses
    if 'Model Classification' not in df.columns:
        df['Model Classification'] = None

    # Iterate through the dataframe
    for index, row in df.iterrows():
        if row['Reference Article Downloaded'] == 'Yes':
            if skip_existing and pd.notna(row['Model Classification']):
                print(f"Skipping existing classification for index {index}")
                continue
            start_time = time.time()
            print(f"Processing: " + row['Reference Article ID'])

            # Create the prompt
            prompt = create_prompt(row, corrected_statements=corrected_statements, two_labels=two_labels)
            
            # Send the prompt and get the response
            response = send_prompt(prompt, model)
            
            # Save the response to the new column
            df.at[index, 'Model Classification'] = response

            if save_intermediate_results and index % save_every == 0:
                df.to_pickle(f"../data/dfs{'/annotated_data' if annotated else ''}{'/two_labels' if two_labels else ''}{'/corrected_statements' if corrected_statements else ''}/{embedding}{'_no_prev_chunking' if no_prev_chunking else ''}/{grobid_model}/ReferenceErrorDetection_data_with_prompt_results_{model.replace(':','.')}_intermed.pkl")
            end_time = time.time()
            print(f"Took {round(end_time - start_time, 2)} seconds")
            print("==================================")
    return df

%run ../scripts/load_df_for_analysis.py

# This function reads the df with the prompt results for the non-annotated data 
# adds the annotation information to the df
# and removes all model classification values for the data rows which were annotated
# so that only these values will be overwritten in the next prompting
def prepare_df2_reprompt_annotated_data(df_annotated, model):
    df2 = sort_df(pd.read_pickle(f"../data/dfs/{embedding}{'_no_prev_chunking' if no_prev_chunking else ''}/{grobid_model}/ReferenceErrorDetection_data_with_prompt_results_{model.replace(':','.')}.pkl"))

    # Remove rows from df_annotated where "Suited for Task" is "Added"
    df_annotated = sort_df(df_annotated[df_annotated['Suited for Task'] != "Added"])

    # Ensure df2 and df_annotated have the same number of rows
    assert len(df2) == len(df_annotated), "Mismatch in the number of rows between df2 and df_annotated"

    # Add missing columns from df_annotated to df2 with empty values
    for column in df_annotated.columns:
        if column not in df2.columns:
            df2[column] = None

    # Add missing columns from df_annotated to df2
    for index, row in df2.iterrows():
        assert row['Reference Article ID'] == df_annotated.iloc[index]['Reference Article ID'], f"Mismatch in Reference Article ID at index {index}"
        assert row['Citing Article ID'] == df_annotated.iloc[index]['Citing Article ID'], f"Mismatch in Citing Article ID at index {index}"

        # Fill the rows with values from df_annotated
        for column in df_annotated.columns:
            if column in df2.columns:
                df2.at[index, column] = df_annotated.iloc[index][column]

    # Reorder the columns in df2 to match the order of columns in df_annotated
    # Keep the columns that df2 has additionally on the very right
    common_columns = df_annotated.columns
    additional_columns = [col for col in df2.columns if col not in common_columns]
    df2 = df2[list(common_columns) + additional_columns]

    # Remove all model classification values for the data rows which were annotated
    for index, row in df2.iterrows():
        if row['Reference Article Downloaded'] == 'Yes' and not pd.isna(row['Citation Function: Main']):
            # Set the model classification to None for the annotated rows
            df2.at[index, 'Model Classification'] = None

    return df2

models = ["llama3.1:70b", "llama3.1:405b", "llama3.3"]
model = models[0]

df2 = prepare_df2_reprompt_annotated_data(df, model)
display(df2.head())
display(df2[df2['Corrected Statement'].notna()].head())

%%time

two_labels = True

df2 = prompting_model(df2, model, skip_existing=True, save_intermediate_results=True, save_every=5, two_labels=two_labels)

df2.to_pickle(f"../data/dfs{'/annotated_data' if annotated else ''}{'/two_labels' if two_labels else ''}{'/corrected_statements' if corrected_statements else ''}/{embedding}{'_no_prev_chunking' if no_prev_chunking else ''}/{grobid_model}/ReferenceErrorDetection_data_with_prompt_results_{model.replace(':','.')}.pkl")
