==> notebooks/first_experiments/07_results_analysis.ipynb <==
# Result Analysis

models = ["gpt-3.5-turbo-0125", "gpt-4-0125-preview", "gpt-4o-2024-05-13", "llama3.1:70b", "llama3.1:405b", "llama3.3", "llama4:scout"]

%run ../scripts/load_df_for_analysis.py

%run ../scripts/df_calculations.py

## Gather Results: Model Comparison

all_results = {
    "te3s": {},
    "te3s_no_prev_chunking": {},
    "te3s_no_prev_chunking_batched": {},
    "te3l": {},
    "te3l_no_prev_chunking": {},
    "te3l_no_prev_chunking_batched": {},
}

def gather_results(model_type, embedding, no_prev_chunking, batched, gpt_model=models[0], annotated=False, corrected_statements=False, two_labels=False):
    global all_results

    df = load_df(model_type, embedding, no_prev_chunking, gpt_model, batched, annotated, corrected_statements, two_labels)
    if 'Suited for Task' in df.columns:
        df = df[df['Suited for Task'] != "Added"]
    df = sort_df(df)
    df = reshape_model_classification(df)
    
    results_all_labels = eval_predictions_all_labels(df)
    results_all_labels_exlude_not_available = eval_predictions_all_labels(df, False)
    results_two_labels = eval_predictions_two_labels(df)
    results_two_labels_exlude_not_available = eval_predictions_two_labels(df, False)

    results = {
        'all_labels': results_all_labels,
        'all_labels_exclude_not_available': results_all_labels_exlude_not_available,
        'two_labels': results_two_labels,
        'two_labels_exclude_not_available': results_two_labels_exlude_not_available
    }

    embedding_string = embedding + ('_no_prev_chunking' if no_prev_chunking else '') + ('_batched' if batched else '')

    if not isinstance(all_results[embedding_string].get(model_type), dict):
        all_results[embedding_string][model_type] = {}
    gpt_model += ('_two_labels' if two_labels else '') + ('_annotated' if annotated else '') + ('_corrected_statements' if corrected_statements else '')
    all_results[embedding_string][model_type][gpt_model] = results
    

### Small GROBID Model
- PDF text extracted with smaller GROBID model 
- full text from TEI document directly fed into index generation 
- Model temperature set to 0 (for top 3 excerpts retrieval and for classification via model prompting)
- Used embedding for index: text-embedding-3-small
- Used model for prompting: gpt-3.5-turbo-0125

#### Small Text Embedding

model_type = "small_model"
embedding = "te3s"
no_prev_chunking = False
batched = False
gather_results(model_type, embedding, no_prev_chunking, batched)

### Full GROBID Model
- PDF text extracted with full GROBID model
- full text from TEI document directly fed into index generation 
- Model temperature set to 0 (for top 3 excerpts retrieval and for classification via model prompting)
- Used embedding for index: text-embedding-3-small
- Used model for prompting: gpt-3.5-turbo-0125

#### Small Text Embedding

model_type = "full_model"
embedding = "te3s"
no_prev_chunking = False
batched = False
gather_results(model_type, embedding, no_prev_chunking, batched)

#### Small Text Embedding - no previous chunking

model_type = "full_model"
embedding = "te3s"
no_prev_chunking = True
batched = False
gather_results(model_type, embedding, no_prev_chunking, batched)

#### Small Text Embedding - no previous chunking - batched

model_type = "full_model"
embedding = "te3s"
no_prev_chunking = True
batched = True
gather_results(model_type, embedding, no_prev_chunking, batched)

#### Large Text Embedding

model_type = "full_model"
embedding = "te3l"
no_prev_chunking = False
batched = False
gather_results(model_type, embedding, no_prev_chunking, batched)

#### Large Text Embedding - no previous chunking

##### GPT 3 Turbo

model_type = "full_model"
embedding = "te3l"
no_prev_chunking = True
batched = False
model = models[0]
gather_results(model_type, embedding, no_prev_chunking, batched)

##### GPT 4

model_type = "full_model"
embedding = "te3l"
no_prev_chunking = True
batched = False
model = models[1]
gather_results(model_type, embedding, no_prev_chunking, batched, model)

##### GPT 4o

model_type = "full_model"
embedding = "te3l"
no_prev_chunking = True
batched = False
model = models[2]
gather_results(model_type, embedding, no_prev_chunking, batched, model)

##### Llama 3.1 (70b)

model_type = "full_model"
embedding = "te3l"
no_prev_chunking = True
batched = False
model = models[3]
gather_results(model_type, embedding, no_prev_chunking, batched, model)

##### Llama 3.1 (405b)

model_type = "full_model"
embedding = "te3l"
no_prev_chunking = True
batched = False
model = models[4]
gather_results(model_type, embedding, no_prev_chunking, batched, model)

##### Llama 3.3 (70b)

model_type = "full_model"
embedding = "te3l"
no_prev_chunking = True
batched = False
model = models[5]
gather_results(model_type, embedding, no_prev_chunking, batched, model)

##### Llama 4 Scout

model_type = "full_model"
embedding = "te3l"
no_prev_chunking = True
batched = False
model = models[6]
gather_results(model_type, embedding, no_prev_chunking, batched, model)

#### Large Text Embedding - no previous chunking - batched

model_type = "full_model"
embedding = "te3l"
no_prev_chunking = True
batched = True
gather_results(model_type, embedding, no_prev_chunking, batched)

### Full GROBID Model + TEI document text refactoring
- PDF text extracted with full GROBID model
- Only text from body of papers (actual content) extracted from TEI documents
- Model temperature set to 0 (for top 3 excerpts retrieval and for classification via model prompting)
- Used embedding for index: text-embedding-3-small
- Used model for prompting: gpt-3.5-turbo-0125

#### Small Text Embedding

model_type = "full_model_texts"
embedding = "te3s"
no_prev_chunking = False
batched = False
gather_results(model_type, embedding, no_prev_chunking, batched)

#### Small Text Embedding - no previous chunking

model_type = "full_model_texts"
embedding = "te3s"
no_prev_chunking = True
batched = False
gather_results(model_type, embedding, no_prev_chunking, batched)

#### Small Text Embedding - no previous chunking - batched

model_type = "full_model_texts"
embedding = "te3s"
no_prev_chunking = True
batched = True
gather_results(model_type, embedding, no_prev_chunking, batched)

#### Large Text Embedding

model_type = "full_model_texts"
embedding = "te3l"
no_prev_chunking = False
batched = False
gather_results(model_type, embedding, no_prev_chunking, batched)

#### Large Text Embedding - no previous chunking

model_type = "full_model_texts"
embedding = "te3l"
no_prev_chunking = True
batched = False
gather_results(model_type, embedding, no_prev_chunking, batched)

#### Large Text Embedding - no previous chunking - batched

model_type = "full_model_texts"
embedding = "te3l"
no_prev_chunking = True
batched = True
gather_results(model_type, embedding, no_prev_chunking, batched)

### Annotated Data

#### Three Labels Classification

##### Llama 3.1 (70b) - Annotated Data

model_type = "full_model"
embedding = "te3l"
no_prev_chunking = True
batched = False

annotated = True

model = models[3]
gather_results(model_type, embedding, no_prev_chunking, batched, model, annotated)

##### Llama 3.1 (70b) - Annotated Data + Corrected Statements

model_type = "full_model"
embedding = "te3l"
no_prev_chunking = True
batched = False

annotated = True
corrected_statements = True

model = models[3]
gather_results(model_type, embedding, no_prev_chunking, batched, model, annotated, corrected_statements)

#### Two Labels (Fully substanitated/Unsubstantiate) Classification

##### Llama 3.1 (70b) - Annotated Data

model_type = "full_model"
embedding = "te3l"
no_prev_chunking = True
batched = False

annotated = True
two_labels = True

model = models[3]
gather_results(model_type, embedding, no_prev_chunking, batched, model, annotated, two_labels=two_labels)

##### Llama 3.1 (70b) - Annotated Data + Corrected Statements

model_type = "full_model"
embedding = "te3l"
no_prev_chunking = True
batched = False

annotated = True
corrected_statements = True
two_labels = True

model = models[3]
gather_results(model_type, embedding, no_prev_chunking, batched, model, annotated, corrected_statements, two_labels=two_labels)

### Save results

import json

with open("../data/all_results.json", "w") as json_file:
    json.dump(all_results, json_file, indent=4)

## Reproduce Table from Paper

results_for_table = all_results['te3l_no_prev_chunking']['full_model']

results_for_table

print("Label Accuracies - Include All Downloaded")
print_table_label_accuracies(calc_label_accuracies(results_for_table))

print("Label Accuracies - Exclude Not Available in Paper")
print_table_label_accuracies(calc_label_accuracies(results_for_table, True))

### Compare to paper results

calc_label_accuracies(results_for_table)

paper_results = {
    'gpt-3.5-turbo-0125': {
        'unsubstantiate': 0.795,
        'partially substantiate': 0.571,
        'fully substantiate': 0.306,
        'overall': 0.540,
    },
    'gpt-4-0125-preview': {
        'unsubstantiate': 0.839,
        'partially substantiate': 0.214,
        'fully substantiate': 0.629,
        'overall': 0.700,
    },
    'gpt-4o-2024-05-13': {
        'unsubstantiate': 0.866,
        'partially substantiate': 0.500,
        'fully substantiate': 0.347,
        'overall': 0.588,
    },
}

def calc_difference_between_results(calculated_results, paper_results):
    differences = {}
    for model, results in paper_results.items():
        if model in calculated_results:
            differences[model] = {
                key: calculated_results[model][key] - value
                for key, value in results.items()
            }
    return differences

calc_difference_between_results(calc_label_accuracies(results_for_table), paper_results)

label_accuracies = calc_label_accuracies(results_for_table, True)
print("Label Accuracies - Exclude Not Available in Paper")
print_table_label_accuracies(label_accuracies)
print("")
print("Paper Results")
print_table_label_accuracies(paper_results)
print("")
print("Differences from paper results to calculated paper results")
print_table_label_accuracies(calc_difference_between_results(label_accuracies, paper_results))
